+++ CHANGELOG.md
@@ -54,6 +54,7 @@ Please also have a look at our
 
 ### Fixed
 
+- Render rules in line and column number order (#1058)
 - Don't render `rgb` colors with percentage values using hex notation (#803)
 - Parse `@font-face` `src` property as comma-delimited list (#790)